requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
colorlog==6.9.0
comtypes==1.4.11
idna==3.10
paho-mqtt==2.1.0
pillow==11.2.1
psutil==7.0.0
pycaw==20240210
pystray==0.19.5
pywin32==310
requests==2.32.4
six==1.17.0
urllib3==2.4.0
rich

